(feature/neeraj-user-activities): break items in chunk
